move ts to new file
