+++ .changeset/thirty-ads-think.md
@@ -1,0 +1,5 @@
+---
+'@microsoft/atlas-site': minor
+---
+
+Add button for easier manual right to left testing.
